% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_models}
\alias{compare_models}
\title{Compare a flat list of models (no nesting) with ANOVA}
\usage{
compare_models(model_list)
}
\description{
Compare a flat list of models (no nesting) with ANOVA
}
\details{
Take care that the underlying model data is equal in the
models that are being compared, otherwise this function will fail to compare
them
}
