% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_linefigures_legend.R
\name{create_linefigures_legend}
\alias{create_linefigures_legend}
\title{This function creates a legend to be used in a multi-grap panel}
\usage{
create_linefigures_legend(DF)
}
\description{
This function creates a legend to be used in a multi-grap panel
}
