% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distro}
\alias{distro}
\title{Toy example of a switch function for educational  puposes
## toy example}
\usage{
distro(n, type, ...)
}
\description{
Toy example of a switch function for educational  puposes
## toy example
}
