% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_lines.R
\name{draw_lines}
\alias{draw_lines}
\title{Create a summary line graph from a single analyte}
\usage{
draw_lines(DF, palette_graph, analyte_x, ...)
}
\arguments{
\item{DF}{A data.frame containing data for a single analyte}

\item{palette_graph}{A palette with 5 colours}
}
\value{
a ggplot2 object and the function writes to
disk in the directory defined by the function
saveInImageDirectory
}
\description{
A function to create a summary line graph from a
dataset based on a single analyte.
}
