% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_lines_panel_6.R
\name{draw_lines_panel_6}
\alias{draw_lines_panel_6}
\title{Create a summary line graph from a single analyte; suitable for multipanelled plots}
\usage{
draw_lines_panel_6(
  DF,
  palette_graph,
  ymax = max(DF$mean_conc * 1.5),
  ymin = min(DF$mean_conc * 1.5)
)
}
\arguments{
\item{DF}{A data.frame containing data for a single analyte}

\item{palette_graph}{A palette with 5 colours}
}
\value{
a ggplot2 object and the function writes to
disk in the directory defined by the function
saveInImageDirectory
}
\description{
A function to create a summary line graph from a
dataset based on a single analyte.
}
