% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_me.R
\name{model_me}
\alias{model_me}
\title{Switch between statistical models by entering a type and formula}
\usage{
model_me(df, formula, type, ...)
}
\arguments{
\item{df}{A dataframe}

\item{formula}{An R formula or model specification, using the expression
form dependent_variable ~ predictors}

\item{type}{One of type 'lm', 'gls' or 'lme'}
}
\description{
This function is a switch function allowing you to explore
all studied models in this package. Currently 'lm', 'gls' and 'lme'
models are implemented
}
\details{
The structure for lme is the most intricate a flexible, the model
formula for the fixed structure of the model can be entered as an R fomula
as stated under 'formula'. The random effects can be added to the model. See
?nlme::lme  for more details (@seealso package {nlme})
}
