% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_boxplot_from_nested_df}
\alias{plot_boxplot_from_nested_df}
\title{Function to plot boxplots from a (nested) dataframe}
\usage{
plot_boxplot_from_nested_df(
  df,
  title,
  var_x,
  var_y,
  xlab = var_x,
  ylab = var_y
)
}
\arguments{
\item{df}{An object of class "data.frame", "tbl", "tbl-df"}
}
\description{
Function to plot boxplots from a (nested) dataframe
}
