% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_histogram_from_nested_df}
\alias{plot_histogram_from_nested_df}
\title{Function to plot histograms from a (nested) dataframe}
\usage{
plot_histogram_from_nested_df(df, var, title)
}
\arguments{
\item{df}{An object of class "data.frame", "tbl", "tbl-df"}
}
\description{
Function to plot histograms from a (nested) dataframe
}
