% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render report of full EDA and statistical analysis salivar_data}
\usage{
render_report()
}
\description{
Render report of full EDA and statistical analysis salivar_data
}
