% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_in_image_directory.R
\name{save_in_image_directory}
\alias{save_in_image_directory}
\title{Function to save a ggplot2 graph to disk}
\usage{
save_in_image_directory(filename, f.height = 9, f.width = 14, ...)
}
\arguments{
\item{filename}{A filename as a full path character string, including extensions}
}
\description{
Function to save a ggplot2 graph to disk
}
