% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_individual.R
\name{theme_individual}
\alias{theme_individual}
\title{Theme for individual line graphs}
\usage{
theme_individual(base_size = 30, base_family = "")
}
\arguments{
\item{base_size}{General setting for size of font}

\item{base_family}{General setting for font type}
}
\value{
a ggplot2 theme
}
\description{
Theme for individual line graphs
}
