% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_panel_2.R
\name{theme_panel_2}
\alias{theme_panel_2}
\title{Theme for multiplot panels line graphs}
\usage{
theme_panel_2(base_size = 30, base_family = "")
}
\arguments{
\item{base_size}{General setting for size of font}

\item{base_family}{General setting for font type}
}
\value{
a ggplot2 theme
}
\description{
Theme for multiplot panels line graphs
}
